requirements.txt
torch>=2.0.1
torchvision>=0.15.1
numpy>=1.24.2
pillow>=9.2.0
pyTelegramBotAPI>=4.10.1
python-dotenv>=0.21.0